+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# `master`
+
+- ✨ New `Threshold` service #115
+
 # `0.86.0`
 
 - 💥 YouLess: replace the `url` setting with the `host` setting